+++ Solutions/Salesforce Service Cloud/Data Connectors/SalesforceServiceCloud_API_FunctionApp.json
@@ -33,7 +33,7 @@
     ],
     "availability": {
         "status": 1,
-        "isPreview": true
+        "isPreview": false
     },
     "permissions": {
         "resourceProvider": [
