+++ Solutions/Salesforce Service Cloud/Data Connectors/azuredeploy_Connector_Salesforce_Service_Cloud_AzureFunction.json
