+++ Solutions/Salesforce Service Cloud/Data/Solution_TSalesforceCloudtemplateSpec.json
@@ -1,0 +1,18 @@
+{
+  "Name": "Salesforce Service Cloud",
+  "Author": "",
+  "Logo": "<img src=\"https://raw.githubusercontent.com/Azure/Azure-Sentinel/master/Logos/salesforce_logo.svg\" width=\"75px\" height=\"75px\">",
+  "Description": "The [Salesforce Service Cloud](https://www.salesforce.com/in/products/service-cloud/overview/) solution for Microsoft Sentinel enables you to ingest Service Cloud events into Microsoft Sentinel.  ",
+  "Data Connectors": [
+    "Data Connectors/SalesforceServiceCloud_API_FunctionApp.json"
+  ],
+  "Parsers": [
+    "Parsers/SalesforceServiceCloud"
+  ],
+
+  "BasePath": "C:\\GitHub\\Azure-Sentinel\\Solutions\\Salesforce Service Cloud",
+  "Version": "2.0.0",
+  "Metadata": "SolutionMetadata.json",
+ "TemplateSpec": true,
+  "Is1PConnector": false
+}
