+++ Solutions/Salesforce Service Cloud/SolutionMetadata.json
@@ -1,0 +1,19 @@
+{
+	"publisherId": "azuresentinel",
+	"offerId": "azure-sentinel-solution-salesforceservicecloud",
+	"firstPublishDate": "2022-05-16",
+	"providers": [""],
+	"categories": {
+		"domains" : [""],
+		"verticals": []
+	},
+	"support": {
+	  "name": "",
+	  "tier": "",
+	  "link": ""
+	}
+}
+
+
+
+
